Update styling and markup for better user experience
Made several changes to the layout and style of the application to make it more user-friendly.
In the CSS, added new selectors for input elements and specific areas for text to provide more aesthetic look.
In the HTML, several textareas and paragraphs are now assigned to the class 'noshow' to hide them for improving overall page cleanliness. These elements may be shown later based on user interaction.