Add drag-and-drop card functionality
This commit introduces a new feature that allows users to reposition cards through drag-and-drop.

In addition to the core drag-and-drop logic, small adjustments in the UI were implemented to allow better user interaction. For instance, textarea boxes in card composer got fixed width for consistency when being dragged. Also, a shadow-box was added to provide visual feedback when the card is being dragged. A grabbed cursor was added to show the card is in 'drag mode'.

Minor changes were as well made to the buttons to improve user interaction, such as redefining the add button class for better CSS targeting and creating a minimalistic cancel button.

In terms of code, "cards" property was added to DOMLogic class to select all cards and a method "cardDrag()" was created and called in app.js to handle card dragging. This method includes mouse events to control the dragging action. The card is appended to a new list when detected that it's dropped over a trello-list.

Minor adjustments to HTML classes and ids to accommodate these new changes.