Refactor card dragging logic in domLogic.js
In the cardDrag function of domLogic.js, made several changes in order to streamline and optimise the dragging process. The variable `elementAfterActualElement` was removed and replaced with `lastCardsRoom`, which is null by default. The `cardsRoom` variable was also removed from the beginning of the logic, to avoid redundant or unnecessary queries. Now, `cardsRoom` and `actualElement` pointer events are handled within nested if blocks, improving readability and efficiency.
In the mouse movements handling part, `cardsRoom` has been replaced with `lastCardsRoom`, reducing the need for direct DOM manipulation. These changes overall are intended to decrease unnecessary computation and enhance performance of the drag operation.